server/requirements.txt

# ============================= requirements.txt ==============================
# Place this one directory up (server/requirements.txt)
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pymupdf==1.24.10
python-docx==1.1.2
# Optional OCR (comment out if build time is tight)
easyocr==1.7.2
torch>=2.2.0
# Optional LLM client
openai>=1.40.0

